backend/requirements.txt
fastapi
uvicorn
python-multipart
boto3
pydantic
python-dotenv
authlib
httpx
itsdangerous
